Moved data store service
