+++ UndoAssessment/UndoAssessment/ViewModels/BaseViewModel.cs
@@ -7,6 +7,7 @@ using Xamarin.Forms;
 
 using UndoAssessment.Models;
 using UndoAssessment.Services;
+using UndoAssessment.Services.Storage;
 
 namespace UndoAssessment.ViewModels
 {
